CHE-5688 Move resolving of servers to creation of machine
